Add file name option to export

+++ src/main/java/net/plarke/bee_statistics/system/JsonExporter.java
@@ -8,13 +8,18 @@ import com.google.gson.GsonBuilder;
 import java.io.FileWriter;
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.FileAlreadyExistsException;
 
 public class JsonExporter {
-    public static void export() {
+    public static void export(String fileName) throws FileAlreadyExistsException {
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
-        //TODO Add option for filename
-        File file = new File("config/hive_logs.json");
+        String filePath = "config/" + fileName + ".json";
+
+        File file = new File(filePath);
+        if (file.isFile()) {
+            throw new FileAlreadyExistsException("Cannot export data to an existing file");
+        }
 
         //TODO Add handling for failed saving
         try (FileWriter writer = new FileWriter(file)) {
